requirements.txt
Flask>=3.0.0
Flask-WTF>=1.2.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.0.0
python-dotenv>=1.0.0
